@@ -246,9 +246,9 @@ abstract class BuildTree extends BuildBit {
     final sb = _buffers[this] = StringBuffer();
     sb.writeln('BuildTree#$hashCode $styleBuilder:');
 
-    const _indent = '  ';
+    const indent = '  ';
     for (final child in children) {
-      sb.write('$_indent${child.toString().replaceAll('\n', '\n$_indent')}\n');
+      sb.write('$indent${child.toString().replaceAll('\n', '\n$indent')}\n');
     }
 
     final str = sb.toString().trimRight();
