@@ -4,6 +4,8 @@ const kCssBackground = 'background';
 const kCssBackgroundColor = 'background-color';
 
 class StyleBgColor {
+  static const kPriorityBoxModel7k5 = 7500;
+
   final WidgetFactory wf;
 
   static final _skipBuilding = Expando<bool>();
@@ -44,7 +46,7 @@ class StyleBgColor {
               widgets;
         },
         onWidgetsIsOptional: true,
-        priority: StyleBorder.kPriorityBoxModel5k + 1,
+        priority: kPriorityBoxModel7k5,
       );
 
   Color? _parseColor(WidgetFactory wf, BuildTree tree) {
