@@ -26,7 +26,7 @@ Widget _paddingInlineSizedBox(double? width) =>
     width != null && width > 0 ? SizedBox(width: width) : widget0;
 
 class StylePadding {
-  static const kPriorityBoxModel3k = 3000;
+  static const kPriorityBoxModel5k = 5000;
 
   final WidgetFactory wf;
 
@@ -69,7 +69,7 @@ class StylePadding {
           ];
         },
         onWidgetsIsOptional: true,
-        priority: kPriorityBoxModel3k,
+        priority: kPriorityBoxModel5k,
       );
 
   Widget? _build(
