@@ -7,64 +7,255 @@ const kCssMinHeight = 'min-height';
 const kCssMinWidth = 'min-width';
 const kCssWidth = 'width';
 
-class DisplayBlockOp extends BuildOp {
-  DisplayBlockOp(WidgetFactory wf)
-      : super(
-          onWidgets: (tree, widgets) => listOrNull(
-            wf
-                .buildColumnPlaceholder(tree, widgets)
-                ?.wrapWith((_, w) => w is CssSizing ? w : CssBlock(child: w)),
-          ),
-          priority: StyleSizing.kPriority7k + 1,
-        );
+extension CssLengthToSizing on CssLength {
+  CssSizingValue? getSizing(HtmlStyle style) {
+    final value = getValue(style);
+    if (value != null) {
+      return CssSizingValue.value(value);
+    }
+
+    switch (unit) {
+      case CssLengthUnit.auto:
+        return const CssSizingValue.auto();
+      case CssLengthUnit.percentage:
+        return CssSizingValue.percentage(number);
+      default:
+        return null;
+    }
+  }
 }
 
 class StyleSizing {
-  static const kPriority7k = 7000;
+  static const k100percent = CssLength(100, CssLengthUnit.percentage);
+  static const kPriorityBoxModel3k = 3000;
 
+  late final BuildOp _opBlock;
+  late final BuildOp _opChild;
+  late final BuildOp _opSizing;
   final WidgetFactory wf;
 
-  static final _treatHeightAsMinHeight = Expando<bool>();
+  static final _elementTree = Expando<BuildTree>();
+  static final _instances = Expando<StyleSizing>();
+  static final _treeIsBlock = Expando<bool>();
+  static final _skipBuilding = Expando<bool>();
 
-  StyleSizing(this.wf);
+  static void registerBlock(WidgetFactory wf, BuildTree tree) {
+    _elementTree[tree.element] = tree;
+    _treeIsBlock[tree] = true;
 
-  BuildOp get buildOp => BuildOp(
-        onTreeFlattening: (tree) {
-          final input = _parse(tree, isDisplayBlock: false);
-          if (input == null) {
-            return;
-          }
+    final instance = StyleSizing._factory(wf);
+    tree
+      ..register(instance._opBlock)
+      ..register(instance._opSizing);
+  }
 
-          WidgetPlaceholder? widget;
-          for (final b in tree.bits) {
-            if (b is WidgetBit) {
-              if (widget != null) {
-                return;
-              }
-              widget = b.child;
-            } else {
-              return;
-            }
-          }
+  static void registerChild(WidgetFactory wf, BuildTree tree) {
+    final parentElement = tree.element.parent;
+    if (parentElement == null || _elementTree[parentElement] == null) {
+      return;
+    }
 
-          widget?.wrapWith((c, w) => _build(c, w, input, tree.styleBuilder));
+    tree.register(StyleSizing._factory(wf)._opChild);
+  }
+
+  static void registerSizing(WidgetFactory wf, BuildTree tree) {
+    _elementTree[tree.element] = tree;
+    tree.register(StyleSizing._factory(wf)._opSizing);
+  }
+
+  factory StyleSizing._factory(WidgetFactory wf) {
+    final existingInstance = _instances[wf];
+    if (existingInstance != null) {
+      return existingInstance;
+    }
+    return _instances[wf] = StyleSizing._(wf);
+  }
+
+  StyleSizing._(this.wf) {
+    _opBlock = const BuildOp(onWidgets: bypass);
+
+    _opChild = BuildOp(
+      onWidgets: buildMinWidthZero,
+      onWidgetsIsOptional: true,
+      // min-width resetter should wrap all other box model widgets
+      priority: StyleMargin.kPriorityBoxModel9k + 1,
+    );
+
+    _opSizing = BuildOp(
+      onTreeFlattening: handleInlineSizing,
+      onWidgets: buildCssSizing,
+      onWidgetsIsOptional: true,
+      priority: StyleSizing.kPriorityBoxModel3k,
+    );
+  }
+
+  Iterable<Widget>? buildCssSizing(
+    BuildTree tree,
+    Iterable<WidgetPlaceholder> widgets,
+  ) {
+    if (_skipBuilding[tree] == true || widgets.isEmpty) {
+      return widgets;
+    }
+
+    final input = _StyleSizingInput.tryParse(tree);
+    if (input == null) {
+      return widgets;
+    }
+
+    return listOrNull(
+      wf
+          .buildColumnPlaceholder(tree, widgets)
+          ?.wrapWith((c, w) => _build(c, w, input, tree.styleBuilder)),
+    );
+  }
+
+  Iterable<Widget>? buildMinWidthZero(
+    BuildTree subTree,
+    Iterable<WidgetPlaceholder> widgets,
+  ) {
+    if (_StyleSizingInput.tryParse(subTree)?.preferredWidth == k100percent) {
+      return widgets;
+    }
+
+    final parentElement = subTree.element.parent;
+    if (parentElement == null) {
+      return widgets;
+    }
+
+    final parentMeta = _elementTree[parentElement];
+    if (parentMeta == null) {
+      return widgets;
+    }
+
+    final parentInput = _StyleSizingInput.tryParse(parentMeta);
+    if (parentInput == null ||
+        (parentInput.minWidth == null && parentInput.preferredWidth == null)) {
+      return widgets;
+    }
+
+    return listOrNull(
+      wf.buildColumnPlaceholder(subTree, widgets)?.wrapWith(
+        (context, child) {
+          final textDirection =
+              subTree.styleBuilder.build(context).textDirection;
+          return _MinWidthZero(
+            textDirection: textDirection,
+            child: child,
+          );
         },
-        onWidgets: (tree, widgets) {
-          final input = _parse(tree, isDisplayBlock: true);
-          if (input == null) {
-            return widgets;
-          }
+      ),
+    );
+  }
 
-          final built = wf.buildColumnPlaceholder(tree, widgets)?.wrapWith(
-                (c, w) => _build(c, w, input, tree.styleBuilder),
-              );
-          return listOrNull(built) ?? widgets;
-        },
-        onWidgetsIsOptional: true,
-        priority: kPriority7k,
-      );
+  static Iterable<Widget>? bypass(
+    BuildTree _,
+    Iterable<WidgetPlaceholder> widgets,
+  ) =>
+      widgets;
 
-  _StyleSizingInput? _parse(BuildTree tree, {required bool isDisplayBlock}) {
+  static void handleInlineSizing(BuildTree tree) {
+    if (_skipBuilding[tree] == true) {
+      return;
+    }
+
+    final input = _StyleSizingInput.tryParse(tree);
+    if (input == null) {
+      return;
+    }
+
+    WidgetPlaceholder? widget;
+    for (final b in tree.bits) {
+      if (b is WidgetBit) {
+        if (widget != null) {
+          return;
+        }
+        widget = b.child;
+      } else {
+        return;
+      }
+    }
+
+    widget?.wrapWith((c, w) => _build(c, w, input, tree.styleBuilder));
+  }
+
+  static void skip(BuildTree tree) {
+    assert(_skipBuilding[tree] != true, 'Built ${tree.element} already');
+    _skipBuilding[tree] = true;
+  }
+
+  static Widget _build(
+    BuildContext context,
+    Widget child,
+    _StyleSizingInput input,
+    HtmlStyleBuilder styleBuilder,
+  ) {
+    if (input.maxHeight == null &&
+        input.maxWidth == null &&
+        input.minHeight == null &&
+        input.minWidth == null &&
+        input.preferredHeight == null &&
+        input.preferredWidth == StyleSizing.k100percent) {
+      if (child is CssBlock) {
+        // deduplicate whenever possible
+        return child;
+      }
+
+      return CssBlock(child: child);
+    }
+
+    final style = styleBuilder.build(context);
+    return CssSizing(
+      maxHeight: input.maxHeight?.getSizing(style),
+      maxWidth: input.maxWidth?.getSizing(style),
+      minHeight: input.minHeight?.getSizing(style),
+      minWidth: input.minWidth?.getSizing(style),
+      preferredAxis: input.preferredAxis,
+      preferredHeight: input.preferredHeight?.getSizing(style),
+      preferredWidth: input.preferredWidth?.getSizing(style),
+      child: child,
+    );
+  }
+}
+
+class _MinWidthZero extends ConstraintsTransformBox {
+  const _MinWidthZero({
+    required Widget child,
+    Key? key,
+    required TextDirection textDirection,
+  }) : super(
+          alignment: textDirection == TextDirection.ltr
+              ? Alignment.topLeft
+              : Alignment.topRight,
+          constraintsTransform: transform,
+          key: key,
+          child: child,
+        );
+
+  static BoxConstraints transform(BoxConstraints bc) =>
+      bc.copyWith(minWidth: 0);
+}
+
+@immutable
+class _StyleSizingInput {
+  final CssLength? maxHeight;
+  final CssLength? maxWidth;
+  final CssLength? minHeight;
+  final CssLength? minWidth;
+  final Axis? preferredAxis;
+  final CssLength? preferredHeight;
+  final CssLength? preferredWidth;
+
+  const _StyleSizingInput({
+    this.maxHeight,
+    this.maxWidth,
+    this.minHeight,
+    this.minWidth,
+    this.preferredAxis,
+    this.preferredHeight,
+    this.preferredWidth,
+  });
+
+  factory _StyleSizingInput.fromTree(BuildTree tree) {
     CssLength? maxHeight;
     CssLength? maxWidth;
     CssLength? minHeight;
@@ -83,12 +274,8 @@ class StyleSizing {
         case kCssHeight:
           final parsedHeight = tryParseCssLength(value);
           if (parsedHeight != null) {
-            if (_treatHeightAsMinHeight[tree] == true) {
-              minHeight = parsedHeight;
-            } else {
-              preferredAxis = Axis.vertical;
-              preferredHeight = parsedHeight;
-            }
+            preferredAxis = Axis.vertical;
+            preferredHeight = parsedHeight;
           }
           break;
         case kCssMaxHeight:
@@ -113,20 +300,11 @@ class StyleSizing {
       }
     }
 
-    if (maxHeight == null &&
-        maxWidth == null &&
-        minHeight == null &&
-        minWidth == null &&
-        preferredHeight == null &&
-        preferredWidth == null) {
-      return null;
-    }
-
-    if (preferredWidth == null && isDisplayBlock) {
+    if (preferredWidth == null && StyleSizing._treeIsBlock[tree] == true) {
       // `display: block` implies a 100% width
       // but it MUST NOT reset width value if specified
       // we need to keep track of block width to calculate contraints correctly
-      preferredWidth = const CssLength(100, CssLengthUnit.percentage);
+      preferredWidth = StyleSizing.k100percent;
       preferredAxis ??= Axis.horizontal;
     }
 
@@ -141,67 +319,23 @@ class StyleSizing {
     );
   }
 
-  static void treatHeightAsMinHeight(BuildTree tree) =>
-      _treatHeightAsMinHeight[tree] = true;
+  static final _treeInput = Expando<_StyleSizingInput>();
 
-  static Widget _build(
-    BuildContext context,
-    Widget child,
-    _StyleSizingInput input,
-    HtmlStyleBuilder styleBuilder,
-  ) {
-    final style = styleBuilder.build(context);
+  static _StyleSizingInput? tryParse(BuildTree tree) {
+    var input = _treeInput[tree];
+    if (input == null) {
+      _treeInput[tree] = input = _StyleSizingInput.fromTree(tree);
+    }
 
-    return CssSizing(
-      maxHeight: _getValue(input.maxHeight, style),
-      maxWidth: _getValue(input.maxWidth, style),
-      minHeight: _getValue(input.minHeight, style),
-      minWidth: _getValue(input.minWidth, style),
-      preferredAxis: input.preferredAxis,
-      preferredHeight: _getValue(input.preferredHeight, style),
-      preferredWidth: _getValue(input.preferredWidth, style),
-      child: child,
-    );
-  }
-
-  static CssSizingValue? _getValue(CssLength? length, HtmlStyle style) {
-    if (length == null) {
+    if (input.maxHeight == null &&
+        input.maxWidth == null &&
+        input.minHeight == null &&
+        input.minWidth == null &&
+        input.preferredHeight == null &&
+        input.preferredWidth == null) {
       return null;
     }
 
-    final value = length.getValue(style);
-    if (value != null) {
-      return CssSizingValue.value(value);
-    }
-
-    switch (length.unit) {
-      case CssLengthUnit.auto:
-        return const CssSizingValue.auto();
-      case CssLengthUnit.percentage:
-        return CssSizingValue.percentage(length.number);
-      default:
-        return null;
-    }
+    return input;
   }
-}
-
-@immutable
-class _StyleSizingInput {
-  final CssLength? maxHeight;
-  final CssLength? maxWidth;
-  final CssLength? minHeight;
-  final CssLength? minWidth;
-  final Axis? preferredAxis;
-  final CssLength? preferredHeight;
-  final CssLength? preferredWidth;
-
-  const _StyleSizingInput({
-    this.maxHeight,
-    this.maxWidth,
-    this.minHeight,
-    this.minWidth,
-    this.preferredAxis,
-    this.preferredHeight,
-    this.preferredWidth,
-  });
 }
