@@ -66,27 +66,40 @@ class StyleTextDecoration {
 HtmlStyle textDecorationColor(HtmlStyle style, Color v) =>
     style.copyWith(textStyle: style.textStyle.copyWith(decorationColor: v));
 
-HtmlStyle textDecorationLine(HtmlStyle style, TextDecorationLine v) {
-  final decoration = style.textStyle.decoration;
-  final lineThough = decoration?.contains(TextDecoration.lineThrough) == true;
-  final overline = decoration?.contains(TextDecoration.overline) == true;
-  final underline = decoration?.contains(TextDecoration.underline) == true;
+HtmlStyle textDecorationLine(HtmlStyle p, TextDecorationLine v) {
+  final parent = p.parent?.textStyle.decoration;
+  final parentOverline = parent?.contains(TextDecoration.overline) == true;
+  final parentLineThrough =
+      parent?.contains(TextDecoration.lineThrough) == true;
+  final parentUnderline = parent?.contains(TextDecoration.underline) == true;
+
+  final current = p.textStyle.decoration;
+  final currentOverline = current?.contains(TextDecoration.overline) == true;
+  final currentLineThrough =
+      current?.contains(TextDecoration.lineThrough) == true;
+  final currentUnderline = current?.contains(TextDecoration.underline) == true;
 
   final list = <TextDecoration>[];
-  if (v.over == true || (overline && v.over != false)) {
+  if (parentOverline || (v.over ?? currentOverline)) {
+    // 1. Honor parent's styling if the line decoration is turned on
+    // 2. Then apply incoing value (if set)
+    // 3. Finally fallback to the current styling
+    //
+    // According to https://developer.mozilla.org/en-US/docs/Web/CSS/text-decoration
+    // > Text decorations are drawn across descendant text elements.
+    // > This means that if an element specifies a text decoration,
+    // > then a child element can't remove the decoration.
     list.add(TextDecoration.overline);
   }
-  if (v.strike == true || (lineThough && v.strike != false)) {
+  if (parentLineThrough || (v.strike ?? currentLineThrough)) {
     list.add(TextDecoration.lineThrough);
   }
-  if (v.under == true || (underline && v.under != false)) {
+  if (parentUnderline || (v.under ?? currentUnderline)) {
     list.add(TextDecoration.underline);
   }
 
-  return style.copyWith(
-    textStyle:
-        style.textStyle.copyWith(decoration: TextDecoration.combine(list)),
-  );
+  final combined = TextDecoration.combine(list);
+  return p.copyWith(textStyle: p.textStyle.copyWith(decoration: combined));
 }
 
 HtmlStyle textDecorationStyle(HtmlStyle style, TextDecorationStyle v) =>
